requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.1
fonttools==4.58.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
